Adds options, credits, and confirm exit menus Creates additional canvases and a confirmation to exit the game
