+++ Assets/Scripts/MenuFunctions/CreditsMenuFunctions.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 968f77538d6071246afe3093e89dcbe8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
