+++ Assets/Scripts/MenuFunctions/OptionsMenuFunctions.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class OptionsMenuFunctions : MonoBehaviour
+{
+    public GameObject mainMenuCanvas;
+    public void MainMenuButtonPressed()
+    {
+        mainMenuCanvas.SetActive(true);
+        gameObject.SetActive(false);
+    }
+}
